refactor: change viewer to example and delete shisen_cpp_viewer_node
